+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unpublished
+- In probe support, renamed `Request::URL` to `Request::Url`
 
 # 0.3.0 (2024-12-12)